requirements.txt
dataclasses==0.8
importlib-resources==5.4.0
numpy==1.19.5
Pillow==8.4.0
pkg_resources==0.0.0
torch==1.9.1+cu102
torchvision==0.10.1
tqdm==4.64.0
typing_extensions==4.1.1
zipp==3.6.0
